moves error component to own file
